[refactor] Move email to admin_theme app
