+++ openwisp_utils/admin_theme/settings.py
@@ -9,3 +9,14 @@ ADMIN_SITE_CLASS = getattr(
 OPENWISP_ADMIN_THEME_LINKS = getattr(settings, 'OPENWISP_ADMIN_THEME_LINKS', [])
 OPENWISP_ADMIN_THEME_JS = getattr(settings, 'OPENWISP_ADMIN_THEME_JS', [])
 ADMIN_DASHBOARD_ENABLED = getattr(settings, 'OPENWISP_ADMIN_DASHBOARD_ENABLED', True)
+
+OPENWISP_EMAIL_TEMPLATE = getattr(
+    settings, 'OPENWISP_EMAIL_TEMPLATE', 'email_template.html',
+)
+
+OPENWISP_EMAIL_LOGO = getattr(
+    settings,
+    'OPENWISP_EMAIL_LOGO',
+    'https://raw.githubusercontent.com/openwisp/openwisp-utils/master/openwisp_utils/'
+    'static/openwisp-utils/images/openwisp-logo.png',
+)
