+++ tests/settings.py
@@ -18,7 +18,6 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     # test project
     'test_project',
-    'openwisp_utils',
     'openwisp_utils.admin_theme',
     'django.contrib.sites',
     # admin
